updates diana api
