+++ diana/diana-key-value/src/main/java/org/jnosql/diana/keyvalue/query/DefaultKeyValueQueryParser.java
@@ -14,7 +14,7 @@
  *   Otavio Santana
  *
  */
-package org.jnosql.diana.kv.query;
+package org.jnosql.diana.keyvalue.query;
 
 import jakarta.nosql.QueryException;
 import jakarta.nosql.Value;
