+++ diana/diana-key-value/src/main/java/org/jnosql/diana/keyvalue/query/GetQueryParser.java
@@ -14,7 +14,8 @@
  *   Otavio Santana
  *
  */
-package org.jnosql.diana.kv.query;
+package org.jnosql.diana.keyvalue.query;
+
 
 import jakarta.nosql.Params;
 import jakarta.nosql.QueryException;
@@ -22,40 +23,41 @@ import jakarta.nosql.ServiceLoaderProvider;
 import jakarta.nosql.Value;
 import jakarta.nosql.keyvalue.BucketManager;
 import jakarta.nosql.keyvalue.KeyValuePreparedStatement;
-import jakarta.nosql.query.RemoveQuery;
-import jakarta.nosql.query.RemoveQuery.RemoveQueryProvider;
+import jakarta.nosql.query.GetQuery;
+import jakarta.nosql.query.GetQuery.GetQueryProvider;
 
-import java.util.Collections;
+import java.util.ArrayList;
 import java.util.List;
 
 import static java.util.stream.Collectors.toList;
 
-final class RemoveQueryParser {
+final class GetQueryParser {
 
-    private final RemoveQueryProvider provider;
+    private final GetQueryProvider provider;
 
-    RemoveQueryParser() {
-        this.provider = ServiceLoaderProvider.get(RemoveQueryProvider.class);
+    GetQueryParser() {
+        this.provider = ServiceLoaderProvider.get(GetQueryProvider.class);
     }
 
     List<Value> query(String query, BucketManager manager) {
 
-        RemoveQuery delQuery = provider.apply(query);
+        GetQuery getQuery = provider.apply(query);
         Params params = Params.newParams();
-        List<Value> values = delQuery.getKeys().stream().map(k -> Values.getValue(k, params)).collect(toList());
+        List<Value> values = getQuery.getKeys().stream().map(k -> Values.getValue(k, params)).collect(toList());
         if (params.isNotEmpty()) {
             throw new QueryException("To run a query with a parameter use a PrepareStatement instead.");
         }
 
         List<Object> keys = values.stream().map(Value::get).collect(toList());
-        manager.remove(keys);
-        return Collections.emptyList();
+        List<Value> result = new ArrayList<>();
+        manager.get(keys).forEach(result::add);
+        return result;
     }
 
     public KeyValuePreparedStatement prepare(String query, BucketManager manager) {
-        RemoveQuery delQuery = provider.apply(query);
+        GetQuery getQuery = provider.apply(query);
         Params params = Params.newParams();
-        List<Value> values = delQuery.getKeys().stream().map(k -> Values.getValue(k, params)).collect(toList());
-        return DefaultKeyValuePreparedStatement.del(values, manager, params, query);
+        List<Value> values = getQuery.getKeys().stream().map(k -> Values.getValue(k, params)).collect(toList());
+        return DefaultKeyValuePreparedStatement.get(values, manager, params, query);
     }
 }
