+++ diana/diana-key-value/src/main/java/org/jnosql/diana/keyvalue/query/Values.java
@@ -14,7 +14,7 @@
  *   Otavio Santana
  *
  */
-package org.jnosql.diana.kv.query;
+package org.jnosql.diana.keyvalue.query;
 
 
 import jakarta.nosql.Params;
