+++ diana/diana-key-value/src/main/resources/META-INF/services/jakarta.nosql.kv.KeyValueEntity$KeyValueEntityProvider
@@ -1,1 +1,1 @@
-org.jnosql.diana.kv.DefaultKeyValueEntityProvider
+org.jnosql.diana.keyvalue.DefaultKeyValueEntityProvider
