+++ diana/diana-key-value/src/test/java/org/jnosql/diana/keyvalue/query/ParamValueTest.java
@@ -14,7 +14,7 @@
  *   Otavio Santana
  *
  */
-package org.jnosql.diana.kv.query;
+package org.jnosql.diana.keyvalue.query;
 
 import jakarta.nosql.Params;
 import jakarta.nosql.QueryException;
