+++ diana/diana-key-value/src/test/java/org/jnosql/diana/keyvalue/query/RemoveQueryParserTest.java
@@ -14,7 +14,7 @@
  *   Otavio Santana
  *
  */
-package org.jnosql.diana.kv.query;
+package org.jnosql.diana.keyvalue.query;
 
 import jakarta.nosql.QueryException;
 import jakarta.nosql.keyvalue.BucketManager;
@@ -32,21 +32,22 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class GetQueryParserTest {
+class RemoveQueryParserTest {
 
-    private GetQueryParser parser = new GetQueryParser();
+
+    private RemoveQueryParser parser = new RemoveQueryParser();
     private BucketManager manager = Mockito.mock(BucketManager.class);
 
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"get \"Diana\""})
+    @ValueSource(strings = {"remove \"Diana\""})
     public void shouldReturnParserQuery1(String query) {
 
         ArgumentCaptor<List<Object>> captor = ArgumentCaptor.forClass(List.class);
 
         parser.query(query, manager);
 
-        Mockito.verify(manager).get(captor.capture());
+        Mockito.verify(manager).remove(captor.capture());
         List<Object> value = captor.getValue();
 
         assertEquals(1, value.size());
@@ -54,14 +55,14 @@ class GetQueryParserTest {
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"get 12"})
+    @ValueSource(strings = {"remove 12"})
     public void shouldReturnParserQuery2(String query) {
 
         ArgumentCaptor<List<Object>> captor = ArgumentCaptor.forClass(List.class);
 
         parser.query(query, manager);
 
-        Mockito.verify(manager).get(captor.capture());
+        Mockito.verify(manager).remove(captor.capture());
         List<Object> value = captor.getValue();
 
         assertEquals(1, value.size());
@@ -69,14 +70,14 @@ class GetQueryParserTest {
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"get {\"Ana\" : \"Sister\", \"Maria\" : \"Mother\"}"})
+    @ValueSource(strings = {"remove {\"Ana\" : \"Sister\", \"Maria\" : \"Mother\"}"})
     public void shouldReturnParserQuery3(String query) {
 
         ArgumentCaptor<List<Object>> captor = ArgumentCaptor.forClass(List.class);
 
         parser.query(query, manager);
 
-        Mockito.verify(manager).get(captor.capture());
+        Mockito.verify(manager).remove(captor.capture());
         List<Object> value = captor.getValue();
 
         assertEquals(1, value.size());
@@ -84,13 +85,13 @@ class GetQueryParserTest {
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"get convert(\"2018-01-10\", java.time.LocalDate)"})
+    @ValueSource(strings = {"remove convert(\"2018-01-10\", java.time.LocalDate)"})
     public void shouldReturnParserQuery4(String query) {
         ArgumentCaptor<List<Object>> captor = ArgumentCaptor.forClass(List.class);
 
         parser.query(query, manager);
 
-        Mockito.verify(manager).get(captor.capture());
+        Mockito.verify(manager).remove(captor.capture());
         List<Object> value = captor.getValue();
 
         assertEquals(1, value.size());
@@ -99,14 +100,14 @@ class GetQueryParserTest {
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"get @id"})
+    @ValueSource(strings = {"remove @id"})
     public void shouldReturnErrorWhenUseParameterInQuery(String query) {
         assertThrows(QueryException.class, () -> parser.query(query, manager));
     }
 
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"get @id"})
+    @ValueSource(strings = {"remove @id"})
     public void shouldReturnErrorWhenDontBindParameters(String query) {
 
         KeyValuePreparedStatement prepare = parser.prepare(query, manager);
@@ -114,7 +115,7 @@ class GetQueryParserTest {
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"get @id"})
+    @ValueSource(strings = {"remove @id"})
     public void shouldExecutePrepareStatement(String query) {
 
         ArgumentCaptor<List<Object>> captor = ArgumentCaptor.forClass(List.class);
@@ -122,11 +123,12 @@ class GetQueryParserTest {
         prepare.bind("id", 10);
         prepare.getResultList();
 
-        Mockito.verify(manager).get(captor.capture());
+        Mockito.verify(manager).remove(captor.capture());
         List<Object> value = captor.getValue();
 
         assertEquals(1, value.size());
 
         MatcherAssert.assertThat(value, Matchers.contains(10));
     }
+
 }
